Conflicts:
	ncep/gov.noaa.nws.ncep.common.dataplugin.mcidas/src/gov/noaa/nws/ncep/common/dataplugin/mcidas/McidasRecord.java
	ncep/gov.noaa.nws.ncep.common.dataplugin.ncpafm/src/gov/noaa/nws/ncep/common/dataplugin/ncpafm/NcPafmRecord.java
	ncep/gov.noaa.nws.ncep.common.dataplugin.ncuair/src/gov/noaa/nws/ncep/common/dataplugin/ncuair/NcUairRecord.java
	ncep/gov.noaa.nws.ncep.edex.plugin.mosaic/src/gov/noaa/nws/ncep/edex/plugin/mosaic/common/MosaicRecord.java

All of these were 'self-resolved' when kdiff3 invoked


Former-commit-id: 0e63553a4efd3be72dd875a176e154bf732d9f13 [formerly fc70139990a46ea885e2e69d04a0e781b9bc0b60] [formerly d7f7c26582596b1ffb17e171ec89e9a296bd663d] [formerly d7f7c26582596b1ffb17e171ec89e9a296bd663d [formerly 092efd6072393b28080df7adcd25b53aa04e7bb3]] [formerly 0e63553a4efd3be72dd875a176e154bf732d9f13 [formerly fc70139990a46ea885e2e69d04a0e781b9bc0b60] [formerly d7f7c26582596b1ffb17e171ec89e9a296bd663d] [formerly d7f7c26582596b1ffb17e171ec89e9a296bd663d [formerly 092efd6072393b28080df7adcd25b53aa04e7bb3]] [formerly 31b3e87ffaa46c35d5d516855da3308d032ba601 [formerly d7f7c26582596b1ffb17e171ec89e9a296bd663d [formerly 092efd6072393b28080df7adcd25b53aa04e7bb3] [formerly 31b3e87ffaa46c35d5d516855da3308d032ba601 [formerly f98354aae305ad227eaf3b36e4393b8aefc80f02]]]]]
Former-commit-id: 31b3e87ffaa46c35d5d516855da3308d032ba601
Former-commit-id: c0b06356858f67ced9b390cf5025f32cbe573080 [formerly fe7ac9662515d973c0e49f11d82ff6a3d1063922] [formerly 53d2822329305c6c3ea099346920acc4ed7e1602] [formerly 048a3c43e8a28b1b07994254af460254ea55cc66 [formerly 606a1f5f70c4344a8a6b5b7da1c351cf7707cda2] [formerly 53d2822329305c6c3ea099346920acc4ed7e1602 [formerly 98447412f8071a7ce3005a219c08f08220ab51a1]]]
Former-commit-id: 8fe8181a1f28813581a1a402cf07d262c33c769f [formerly 3940ab38fea505f445ec4019600081324e15e68d] [formerly d040b14742d82fb2e6f50c8c16e2e7295687bf08 [formerly 7ee04eb56006a18226fb322e17ca66f1173ef4bd]]
Former-commit-id: 5576c8f65a9bb6c15ab79ac134243f83cc00f1be [formerly d040b14742d82fb2e6f50c8c16e2e7295687bf08]
Former-commit-id: 2f6ae50479cffb5c4a51c950ce05edc0b322a249

+++ ncep/gov.noaa.nws.ncep.edex.common/src/gov/noaa/nws/ncep/edex/common/sounding/NcSoundingStnInfo.java
@@ -41,9 +41,11 @@ public class NcSoundingStnInfo implements  ISerializableObject{
 	@DynamicSerializeElement
     private float stationElevation;
     @DynamicSerializeElement
-    private float	stationLatitude;
+    //private float	stationLatitude;
+    private double	stationLatitude;
     @DynamicSerializeElement
-    private float	stationLongitude;
+    //private float	stationLongitude;
+    private double	stationLongitude;
     @DynamicSerializeElement
     private String stnId;
     @DynamicSerializeElement
@@ -67,16 +69,16 @@ public class NcSoundingStnInfo implements  ISerializableObject{
 	public void setStationElevation(float stationElevation) {
 		this.stationElevation = stationElevation;
 	}
-	public float getStationLatitude() {
+	public double getStationLatitude() {
 		return stationLatitude;
 	}
-	public void setStationLatitude(float stationLatitude) {
+	public void setStationLatitude(double stationLatitude) {
 		this.stationLatitude = stationLatitude;
 	}
-	public float getStationLongitude() {
+	public double getStationLongitude() {
 		return stationLongitude;
 	}
-	public void setStationLongitude(float stationLongitude) {
+	public void setStationLongitude(double stationLongitude) {
 		this.stationLongitude = stationLongitude;
 	}
 	public String getStnId() {
